LPD-59495 Move modal folder outside components
