+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/FDSPropsTransformer/actions/importStructureAction.ts
@@ -5,7 +5,7 @@
 
 import {openModal} from 'frontend-js-components-web';
 
-import ImportStructureModalContent from '../../components/modal/ImportStructureModalContent';
+import ImportStructureModalContent from '../../modal/ImportStructureModalContent';
 
 export default function importStructureAction(
 	importURL: string,
